requirements.txt
torch>=2.1.0
diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors
scipy
numpy
Pillow
tqdm
pyyaml
matplotlib
seaborn
pandas

# Metrics
torchmetrics[image]>=1.2.0
open_clip_torch>=2.24.0
lpips>=0.1.4
clean-fid>=0.1.35

# Optional: profiling
tensorboard
